feat: extract Client into apps.clients app — zero-data-migration using SeparateDatabaseAndState
- New app apps/clients with model, serializers, views, admin, migrations
- Client model keeps db_table='register_client' (no physical table change)
- 3 state-only migrations: clients/0001_initial, register/0038_remove_client_state, agenda/0007_update_client_fk_to_clients
- register.0038 depends on agenda.0006 to guarantee correct replay order in fresh DBs
- Removed register/serializers_clients.py and register/client_views.py
- All imports updated across tests and management commands
- 79 tests passing
