+++ backend/apps/agenda/tests.py
@@ -3,7 +3,8 @@ from django.utils import timezone
 import pytest
 from rest_framework.test import APIClient
 
-from apps.register.models import Professional, Client
+from apps.clients.models import Client
+from apps.register.models import Professional
 from .models import Appointment
