+++ backend/apps/register/tests/test_client_edit_during_ongoing.py
@@ -2,7 +2,8 @@ import pytest
 from django.utils import timezone
 from rest_framework.test import APIClient
 
-from apps.register.models import Professional, Client
+from apps.clients.models import Client
+from apps.register.models import Professional
 from apps.agenda.models import Appointment
